FIXED WINDOWS BUILD ISSUES
