+++ pyproject.toml
@@ -1,3 +1,3 @@
 [build-system]
-requires = ["setuptools>=45", "wheel", "cffi>=1.15.0"]
+requires = ["setuptools>=61", "wheel", "cffi>=1.15.0"]
 build-backend = "setuptools.build_meta"
