+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = trajectorize
-version = 0.0.4
+version = 0.0.6
 description = KSP Trajectory Optimization
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -31,7 +31,7 @@ package_dir =
 python_requires = >=3.8
 install_requires =
     cffi >= 1.15.0
-    numpy >= 1.24.0
+    numpy >= 1.23.0
     matplotlib >= 3.6.0
 setup_requires =
     cffi >= 1.15.0
